{{cookiecutter.project_slug}}/requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==10.3.0  # https://github.com/python-pillow/Pillow
{%- if cookiecutter.frontend_pipeline == 'Django Compressor' %}
{%- if cookiecutter.windows == 'y' and cookiecutter.use_docker == 'n' %}
rcssmin==1.1.2 --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
{%- else %}
rcssmin==1.1.2  # https://github.com/ndparker/rcssmin
{%- endif %}
{%- endif %}
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
{%- if cookiecutter.use_whitenoise == 'y' %}
whitenoise==6.7.0  # https://github.com/evansd/whitenoise
{%- endif %}
redis==5.0.7  # https://github.com/redis/redis-py
{%- if cookiecutter.use_docker == "y" or cookiecutter.windows == "n" %}
hiredis==2.3.2  # https://github.com/redis/hiredis-py
{%- endif %}
{%- if cookiecutter.use_celery == "y" %}
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.6.0  # https://github.com/celery/django-celery-beat
{%- if cookiecutter.use_docker == 'y' %}
flower==2.0.1  # https://github.com/mher/flower
{%- endif %}
{%- endif %}
{%- if cookiecutter.use_async == 'y' %}
uvicorn[standard]==0.30.1  # https://github.com/encode/uvicorn
uvicorn-worker==0.2.0  # https://github.com/Kludex/uvicorn-worker
{%- endif %}

# Django
# ------------------------------------------------------------------------------
django==4.2.13  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.5.1  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==0.63.3  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.2  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.2  # https://github.com/django-crispy-forms/crispy-bootstrap5
{%- if cookiecutter.frontend_pipeline == 'Django Compressor' %}
django-compressor==4.5  # https://github.com/django-compressor/django-compressor
{%- endif %}
django-redis==5.4.0  # https://github.com/jazzband/django-redis
{%- if cookiecutter.use_drf == 'y' %}
# Django REST Framework
djangorestframework==3.15.2  # https://github.com/encode/django-rest-framework
django-cors-headers==4.4.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.27.2  # https://github.com/tfranzel/drf-spectacular
{%- endif %}
{%- if cookiecutter.frontend_pipeline == 'Webpack' %}
django-webpack-loader==3.1.0  # https://github.com/django-webpack/django-webpack-loader
{%- endif %}

django-extensions
djangorestframework-datatables-editor
django-htmx
django-taggit
django-widget-tweaks
django-filter
django-render-block
pandas
requests
lxml
beautifulsoup4
